Catches ValueError raised when checkpoint is lost.
This happens when the eval job is standalone and is running more than 1 eval. By the time the first few evals finish, the checkpoint for the step being processed may be deleted (because checkpoints to be kept might be low and training might be too fast). When this happens the checkpoint is not found and a ValueError is raised. The top checkpoint then needs to be recalculate which is achieved in this case using `reload`.

PiperOrigin-RevId: 648372277