Allow `ord` to be not `None` in `kd.metrics.Norm`
Currently this fails:
```
kd.metrics.Norm(
  tensor="params.my_network", axis=None, ord=1
)
```

This is because `axis=None` and `ord=None` is a special case in Numpy that
calls `.ravel()` on tensor before calculating the L2 norm.

Adding `axis` won't help, since when passing `ord` the tensor can only be 1-D (for vector norms) and 2-D (for matrix norms). For arbitrary tensors, you always need `.ravel()`

This patch allows to use `ord=1`.

PiperOrigin-RevId: 663767678