Fix Norm with axis=None
We see this pattern a lot:
```
      "grad_norm": kd.metrics.TreeReduce(
          metric=kd.metrics.Norm(tensor="grads", axis=None)
      ),
```

However, this pattern is faulty. This will compute the norm of each tensor in each of the leaves. Then, this will average the norms over all tensors.

Instead, we want to be able to compute the norms as if all parameters were concatenated into a single vector, and compute the norm of that.

This CL fixes this, allowing for a new `aggregation_type` set to `concat`.

It preserves the current default, and writes a warning on an erroneous behavior.

PiperOrigin-RevId: 672880694