(1) The previous method for passing attributes from the linen module to the nnx module was not working well with kauldron's way of choosing a model_method in the evaluators.
The `nnx_wrapper.py` is updated to dynamically expose all methods of the wrapped NNX module as methods of the Linen wrapper.

(2) implement a reseed method for the wrapper's inference. The previous method did not propagate rngs updates through multiple eval steps. The current method reseeds the rng variables based on the rng input of the linen module. It does not use nnx.reseed because nnx.reseed does not work for modules that share rng keys between different sub-objects.

PiperOrigin-RevId: 819061677